requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.24.3
streamlit==1.28.1
pillow==10.0.1
matplotlib==3.7.2
scikit-learn==1.3.0
pandas==2.0.3
tensorflow==2.15.0
